+++ web/src/App.tsx
@@ -9,6 +9,7 @@ import { Documentation } from "./pages/Documentation";
 import { DocumentationGuideView } from "./pages/DocumentationGuideView";
 import { ReleaseNotesList } from "./pages/ReleaseNotesList";
 import { ReleaseNotesVersion } from "./pages/ReleaseNotesVersion";
+import { PortalConfiguration } from "./pages/PortalConfiguration";
 import { Callback } from "./pages/Callback";
 import { Login } from "./pages/Login";
 
@@ -40,6 +41,7 @@ export default function App() {
           }
         >
           <Route index element={<DeviceList />} />
+          <Route path="portal-configuration" element={<PortalConfiguration />} />
           <Route path="documentation" element={<Documentation />} />
           <Route path="documentation/release-notes" element={<ReleaseNotesList />} />
           <Route